FIR2IR: refactor adapter generation
